Load nested Typesafe configs more correctly.
When using the ConfigObject interface of Typesafe Config, the
key provided during traversal is a relative path, not the
path from the root. Also, you cannot use the full path to
refer to particular items within a list.

Instead here, the code uses a traversal of the Config using
entrySet, and then a more specialised conversion between the
list types.