Add ability to use a custom generated Typesafe Config
The current implementation is limited to just calling
ConfigFactory.load. This is likely fine for a lot of use-cases but it
would be a good idea to allow custom Typesafe Config instances to be
passed in and used. This change keeps the existing behavior intact but
exposes an overloaded config method that accepts an instance of an
existing Typesafe Config object.